pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "multiagent-codebase-feedback-assistant"
version = "0.1.0"
description = "AI-based multi-agent system for analyzing Python codebases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.17.0",
    "fastapi>=0.119.0",
    "uvicorn>=0.31.0",
    "litellm>=1.78.5",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.3.0",
    "docker>=7.1.0",
    "rich>=14.2.0",
]

[project.scripts]
codebase-analyzer = "cli.main:main"

[tool.setuptools.packages.find]
include = ["agents*", "cli*", "common*", "tools*", "typings*"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 119
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
mypy_path = "typings"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["litellm.*"]
ignore_missing_imports = true
